clipped_region: make fradii private
